Fix traversal order of `traverseMapWithKey`.
The current definition of `traverseMapWithKey` does not traverse keys in
ascending order:

```hs
> import qualified Data.IntMap as IM
> import qualified Data.IntMap.NonEmpty.Internal as NEIM
> kvs = [(i, i) | i <- [-2 .. 2]]
> () <$ IM.traverseWithKey (curry print) (IM.fromList kvs)
(-2,-2)
(-1,-1)
(0,0)
(1,1)
(2,2)
> () <$ NEIM.traverseMapWithKey (curry print) (IM.fromList kvs)
(-2,-2)
(-1,-1)
(2,2)
(1,1)
(0,0)
```

As a result, traversal functions for `Data.IntMap.NonEmpty` also do not
traverse keys in ascending order.

This commit redefines `traverseMapWithKey` to be a simple synonym of
`Data.IntMap.traverseWithKey`.

With this change, the test suite passes for all versions of containers
from `0.6.3.1` up to `0.8` inclusively:

- `0.6.3.1`
- `0.6.4.1`
- `0.6.5.1`
- `0.6.6`
- `0.6.7`
- `0.6.8`
- `0.7`
- `0.8`