Support `containers-0.8`. (#18)
* Make `combineEq` match bias of `Set.from{Asc,Desc}List`.

This commit provides a workaround for a change in the behaviour of
`Set.from{Asc,Desc}List` w.r.t. duplicate elements, when upgrading
from `containers-0.7` to `containers-0.8`.

We can use the `K` type from `Tests.Util` to demonstrate this.

With `containers-0.7`, `Set.from{Asc,Desc}List` chooses the **_first_**
occurrence of any duplicated element:

```hs
$ cabal repl nonempty-containers-test --constraint=containers==0.7
> import qualified Data.Set as Set
> import Tests.Util
> Set.fromAscList [K 'a' 1, K 'a' 2]
fromList [K {getKX = 'a', getKY = 1}]
> Set.fromDescList [K 'a' 1, K 'a' 2]
fromList [K {getKX = 'a', getKY = 1}]
```

With `containers-0.8`, `Set.from{Asc,Desc}List` chooses the **_last_**
occurrence of any duplicated element:

```hs
$ cabal repl nonempty-containers-test --constraint=containers==0.8
> import qualified Data.Set as Set
> import Tests.Util
> Set.fromAscList [K 'a' 1, K 'a' 2]
fromList [K {getKX = 'a', getKY = 2}]
> Set.fromDescList [K 'a' 1, K 'a' 2]
fromList [K {getKX = 'a', getKY = 2}]
```

This commit adjusts the bias of `Data.Set.NonEmpty.combineEq` to match
the bias of `Set.from{Asc,Desc}List` in the version of `containers`
being linked against.

* Uncomment tests for `IntMap` traversal functions.

* Fix traversal order of `traverseMapWithKey`.

The current definition of `traverseMapWithKey` does not traverse keys in
ascending order:

```hs
> import qualified Data.IntMap as IM
> import qualified Data.IntMap.NonEmpty.Internal as NEIM
> kvs = [(i, i) | i <- [-2 .. 2]]
> () <$ IM.traverseWithKey (curry print) (IM.fromList kvs)
(-2,-2)
(-1,-1)
(0,0)
(1,1)
(2,2)
> () <$ NEIM.traverseMapWithKey (curry print) (IM.fromList kvs)
(-2,-2)
(-1,-1)
(2,2)
(1,1)
(0,0)
```

As a result, traversal functions for `Data.IntMap.NonEmpty` also do not
traverse keys in ascending order.

This commit redefines `traverseMapWithKey` to be a simple synonym of
`Data.IntMap.traverseWithKey`.

With this change, the test suite passes for all versions of containers
from `0.6.3.1` up to `0.8` inclusively:

- `0.6.3.1`
- `0.6.4.1`
- `0.6.5.1`
- `0.6.6`
- `0.6.7`
- `0.6.8`
- `0.7`
- `0.8`

* Inline definition of `traverseMapWithKey`.

Since `traverseMapWithKey` is now a synonym of `IntMap.traverseWithKey`,
we can inline its definition at all call sites.

* Update bounds for `containers`.

The lower bound of `0.6.3.1` is the earliest version for which all tests
pass. This version of `containers` was released on 2020-07-14.

This commit excludes older versions of `containers` up to and including
version `0.6.2.1`, which was released on 2019-06-24.

* Remove compatibility code for `containers < 0.5.11`.

If the package itself is only buildable for `containers >= 0.6.3.1`,
then this code is no longer reachable, and thus can be removed.

* Inline compatibility functions.

These functions are now just synonyms, so they can be safely inlined.